Getting Ready to Animate
Just added a folder where we'll add animation files.  Also put a controller in for Shadow.
Also added music to the camera in our test scene.  To add background music, just drag the sound file onto the camera, check Loop, Play on Awake, and also make sure it's set to 2D, and you should be good!